fix: badge select selector doesnt refreshes
the reason for the problem is that the $state.context state is refreshed
on every time a send command is executed. since we're passsing a new
option array on state changes the component refreshes each time which
causes the selector to unselect.

the solution i settled with is to extract the mapping of the options
outside of the jsx rendering. this way it only happens once, namely,
when the account is selected and nft resources have been fetched